Move ConsoleDynamicMethodFixture to MultiFunctionApplicationHelpers
