+++ tests/Agent/IntegrationTests/IntegrationTests/WCF/WCFServiceTests.cs
@@ -2,14 +2,14 @@
 // SPDX-License-Identifier: Apache-2.0
 
 
-using System;
-using System.Collections.Generic;
-using System.Linq;
+using MultiFunctionApplicationHelpers;
 using NewRelic.Agent.IntegrationTestHelpers;
 using NewRelic.Agent.IntegrationTests.Shared.Wcf;
 using NewRelic.Agent.IntegrationTestHelpers.Models;
-using NewRelic.Agent.IntegrationTests.RemoteServiceFixtures;
 using NewRelic.Testing.Assertions;
+using System;
+using System.Collections.Generic;
+using System.Linq;
 using Xunit;
 using Xunit.Abstractions;
