+++ tests/Agent/IntegrationTests/SharedApplications/ConsoleMultiFunctionApplicationCore/ConsoleMultiFunctionApplicationCore.csproj
@@ -6,7 +6,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="NewRelic.Agent.Api" Version="8.19.353" />
+    <PackageReference Include="NewRelic.Agent.Api" Version="8.36.0" />
   </ItemGroup>
 
   <ItemGroup>
